Removing useless lines

+++ src/StandardsSearchIndexer/Sfa.Eds.Standards.Indexer.AzureWorkerRole/Services/StandardService.cs
@@ -25,27 +25,19 @@ namespace Sfa.Eds.Standards.Indexer.AzureWorkerRole.Services
 
         public async void CreateScheduledIndex(DateTime scheduledRefreshDateTime)
         {
-            Trace.TraceInformation(counter++ + "Getting index alias");
             var indexAlias = GetIndexAlias();
 
-            Trace.TraceInformation(counter++ + "Getting new index name");
             var newIndexName = GetIndexNameAndDateExtension(indexAlias, scheduledRefreshDateTime);
 
-            Trace.TraceInformation(counter++ + "Setting node");
-            //var node = new Uri(ConfigurationManager.AppSettings["ServerUri"]);
-            var node = new Uri("http://104.45.94.2:9200/");
+            var node = new Uri(ConfigurationManager.AppSettings["ServerUri"]);
 
             var connectionSettings = new ConnectionSettings(node, newIndexName);
 
-            Trace.TraceInformation(counter++ + "New Elasticsearch client");
             _client = new ElasticClient(connectionSettings);
 
-            Trace.TraceInformation(counter++ + "Create index if doesn't exists");
             var existingPreviousIndex = CreateIndex(newIndexName);
-            Trace.TraceInformation(counter++ + "Checking if index exists");
             if (existingPreviousIndex) return;
 
-            Trace.TraceInformation(counter++ + "Indexing pdfs");
             await IndexStandards(newIndexName);
 
             PauseWhileIndexingIsBeingRun();
@@ -58,8 +50,7 @@ namespace Sfa.Eds.Standards.Indexer.AzureWorkerRole.Services
 
         private static string GetIndexAlias()
         {
-            //return ConfigurationManager.AppSettings["StandardIndexesAlias"];
-            return "standardIndexAlias";
+            return ConfigurationManager.AppSettings["StandardIndexesAlias"];
         }
 
         private static string GetIndexNameAndDateExtension(string indexAlias, DateTime dateTime)
@@ -69,9 +60,6 @@ namespace Sfa.Eds.Standards.Indexer.AzureWorkerRole.Services
 
         private static bool CreateIndex(string indexName)
         {
-            //indexName = "patatuela";
-            Trace.TraceInformation("Creating index with name: " + indexName);
-
             var indexExistsResponse = _client.IndexExists(indexName);// IndexExists(i => i.Index(indexName));
 
             //If it already exists and is empty, let's delete it.
@@ -83,11 +71,11 @@ namespace Sfa.Eds.Standards.Indexer.AzureWorkerRole.Services
                     return c;
                 });
 
-                //if (totalResults.Count == 0)
-                //{
+                if (totalResults.Count == 0)
+                {
                     _client.DeleteIndex(indexName);
                     indexExistsResponse = _client.IndexExists(indexName);
-                //}
+                }
             }
 
             //create index
@@ -107,15 +95,11 @@ namespace Sfa.Eds.Standards.Indexer.AzureWorkerRole.Services
 
         static public async Task UploadStandardsContentToAzure(List<JsonMetadataObject> standardList)
         {
-            Trace.TraceInformation(counter++ + "EMPIEZA updateo de pdfs");
-
             foreach (var standard in standardList)
             {
                 //await UploadStandardJson(standard);
                 await UploadStandardPdf(standard);
             }
-            Trace.TraceInformation(counter++ + "ACABADO updateo de pdfs");
-
         }
 
         static public async Task UploadStandardJson(JsonMetadataObject standard)
@@ -132,8 +116,6 @@ namespace Sfa.Eds.Standards.Indexer.AzureWorkerRole.Services
 
         static public async Task<List<JsonMetadataObject>> GetStandardsFromAzure()
         {
-            Trace.TraceInformation(counter++ + "Pillando Standards de Azure");
-
             var standardsList = new List<JsonMetadataObject>();
 
             BlobStorageHelper bsh = new BlobStorageHelper();
@@ -141,15 +123,11 @@ namespace Sfa.Eds.Standards.Indexer.AzureWorkerRole.Services
 
             standardsList = standardsList.OrderBy(s => s.Id).ToList();
             
-            Trace.TraceInformation(counter++ + "Standards de Azure ya pillados: " + standardsList.Count);
-
             return standardsList;
         }
 
         private static async Task IndexStandardPdfs(string indexName, List<JsonMetadataObject> standards)
         {
-            Trace.TraceInformation(counter++ + "Indexando documentos");
-
             //index the items
             foreach (var standard in standards)
             {
@@ -159,29 +137,20 @@ namespace Sfa.Eds.Standards.Indexer.AzureWorkerRole.Services
                     .Index(indexName)
                     .Id(doc.StandardId));
             }
-
-            Trace.TraceInformation(counter++ + "Supuestamente indexado");
         }
 
         private static async Task IndexStandards(string newIndexName)
         {
-            Trace.TraceInformation(counter++ + "Vamos a pillar la info de Azure");
-
             var standards = await GetStandardsFromAzure();
-
-            Trace.TraceInformation(counter++ + "Ahora toca updatear los pdfs a azure");
 
             await UploadStandardsContentToAzure(standards);
 
-            Trace.TraceInformation(counter++ + "VAMOS A INDEXAR ALGO");
-            
             try
             {
                 await IndexStandardPdfs(newIndexName, standards);
             }
             catch (Exception e)
             {
-                Trace.TraceError(counter++ + "ERROR!!!! -> " + e.Message);
                 var error = e;
             }
         }
